Please review
@omgmakeme

I have the structure but I wasn't able to get it to work. I spent 40 minutes in office hours and ultimately we figured out that sqlite doesn't allow for columns to be dynamically called via placeholders. Not sure how to modify a row based on user input without creating a method for each input option. This however would not be very DRY at all... Also, you can ignore the category table. I was planning on linking it as a foreign key to the success table, but wasn't able to get to it because I was trying to fix all the other issues.

Can you close your review with an explicit assessment of where you think I am going into phase1? Please feel comfortable to spare my feelings, and give a clear opinion. Thanks!